Save Fibaro device name in hapCharacteristic.SerialNumber to identify device
Fibaro saves the device ID in the default device name. In HomeKit this name is usually replaced by a more speaking name, often same name for same type of device in different rooms, like "ceiling lamp". Renamed to same same names the relation to the respective Fibaro device is only maintained by the room the device is in. If the room is messed up by mistake, it is hard to figure out the related Fibaro device again. Saving the Fibaro device name in the SerialNumber solves this issue. This SerialNumber is visible in the HomeKit device properties. (I hope I figured out the right change for that.)